voting: extract form-factor-neutral flows and shared widgets
Poll card/status badge/message, receipt card, results header/cards, and
the detail summary/meta widgets move to shared widget files. The detail,
review, results, and submission-confirmation screens' state machines move
to builder-based flow widgets so the upcoming mobile screens drive the
exact same logic; the desktop screens become thin shells over them.
No behavior change.
