+++ lib/src/features/voting/widgets/voting_receipt_widgets.dart
@@ -1,0 +1,72 @@
+import 'package:flutter/material.dart';
+
+import '../../../core/theme/app_theme.dart';
+
+/// Label/value receipt rows on a subtle raised card, shared by the desktop
+/// and mobile submission confirmation screens.
+class VotingReceiptCard extends StatelessWidget {
+  const VotingReceiptCard({super.key, required this.rows});
+
+  final List<VotingReceiptRow> rows;
+
+  @override
+  Widget build(BuildContext context) {
+    final colors = context.colors;
+    return Container(
+      decoration: BoxDecoration(
+        color: colors.background.raised.withValues(alpha: 0.78),
+        borderRadius: BorderRadius.circular(AppRadii.medium),
+        border: Border.all(color: colors.border.subtle),
+      ),
+      child: Column(
+        children: [
+          for (var i = 0; i < rows.length; i++) ...[
+            if (i > 0) Divider(height: 1, color: colors.border.subtle),
+            Padding(
+              padding: const EdgeInsets.symmetric(
+                horizontal: AppSpacing.sm,
+                vertical: AppSpacing.s,
+              ),
+              child: rows[i],
+            ),
+          ],
+        ],
+      ),
+    );
+  }
+}
+
+class VotingReceiptRow extends StatelessWidget {
+  const VotingReceiptRow({super.key, required this.label, required this.value});
+
+  final String label;
+  final String value;
+
+  @override
+  Widget build(BuildContext context) {
+    final colors = context.colors;
+    return Row(
+      children: [
+        Expanded(
+          child: Text(
+            label,
+            style: AppTypography.bodyMedium.copyWith(
+              color: colors.text.secondary,
+            ),
+          ),
+        ),
+        const SizedBox(width: AppSpacing.sm),
+        Flexible(
+          child: Text(
+            value,
+            textAlign: TextAlign.right,
+            overflow: TextOverflow.ellipsis,
+            style: AppTypography.bodyMediumStrong.copyWith(
+              color: colors.text.accent,
+            ),
+          ),
+        ),
+      ],
+    );
+  }
+}
